requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]